fix: min heap
